prep.sh now unpacks helper only if the current helper is of newer version. Change migrate cache to /data/local/tmp/migrate_cache. option to see last logs.

+++ app/src/main/assets/prep.sh
@@ -1,5 +1,11 @@
 #!sbin/sh
 
+# parameters:
+
+# TEMP_DIR_NAME
+# version
+# timeStamp
+
 OUTFD="/dev/null"
 
 for FD in `ls /proc/$$/fd`; do
@@ -47,9 +53,9 @@ else
 	exit 1
 fi
 
-if [ -e /tmp/package-data ]
+if [ -e /tmp/package-data.txt ]
 then
-	pd=/tmp/package-data
+	pd=/tmp/package-data.txt
 	while read -r line || [[ -n "$line" ]]; do
 		key=$(echo "$line" | cut -d ' ' -f1)
 		val=$(echo "$line" | cut -d ' ' -f2)
@@ -145,10 +151,43 @@ mkdir -p /system/app/
 mkdir -p /system/priv-app/
 mkdir -p /data/app/
 mkdir -p /data/data/
-rm -r /system/app/PermissionFixer/
-rm -r /data/data/balti.migratehelper/
 
-mkdir -p /data/balti.migrate
-mv /tmp/package-data* /data/balti.migrate/package-data*
+helper_apk_dir=/system/app/MigrateHelper
+
+if [ -e ${helper_apk_dir}/MigrateHelper.apk ]; then
+
+    if [ -e ${helper_apk_dir}/v ]; then
+        last_helper_version="$(cat ${helper_apk_dir}/v)"
+        echo "ui_print Last helper version: $last_helper_version" >> /proc/self/fd/$OUTFD;
+    else
+        last_helper_version="0"
+    fi
+
+    if [ ${last_helper_version} -lt $2 ]; then
+        echo "ui_print Upgrading helper." >> /proc/self/fd/$OUTFD;
+        rm -r ${helper_apk_dir}
+        rm -r /data/data/balti.migratehelper/
+        mkdir -p ${helper_apk_dir}
+        mv -f /tmp/helper.apk ${helper_apk_dir}/MigrateHelper.apk
+        echo "$2" > ${helper_apk_dir}/v
+    else
+        echo "ui_print Helper already present. Skipping helper injection." >> /proc/self/fd/$OUTFD;
+    fi
+else
+    echo "ui_print Injecting helper." >> /proc/self/fd/$OUTFD;
+    rm -r ${helper_apk_dir}
+    rm -r /data/data/balti.migratehelper/
+    mkdir -p ${helper_apk_dir}
+    mv -f /tmp/helper.apk ${helper_apk_dir}/MigrateHelper.apk
+    touch ${helper_apk_dir}/v
+    echo "$2" > ${helper_apk_dir}/v
+fi
+
+if [ -e $1 ] && [ -f $1 ]; then
+    mv $1 $1.bak
+fi
+
+mkdir -p $1
+cp /tmp/package-data.txt $1/package-data${3}.txt
 
 rm /tmp/prep.sh
